+++ services/dy-jupyter/docker/state_puller.py
@@ -14,35 +14,48 @@ from enum import IntEnum
 from pathlib import Path
 
 from simcore_sdk.node_data import data_manager
-from simcore_sdk.node_ports import exceptions
 
 log = logging.getLogger(__file__ if __name__ == "__main__" else __name__)
 logging.basicConfig(level=logging.INFO)
+
 
 class ExitCode(IntEnum):
     SUCCESS = 0
     FAIL = 1
 
-def main(args = None) -> int:
+
+async def pull_file_if_exists(path: Path) -> None:
+    """
+    If the path already exist in storage pull it. Otherwise it is assumed
+    this is the first time the service starts.
+
+    In each and every other case an error is raised and logged
+    """
+    if not await data_manager.is_file_present_in_storage(path):
+        log.info("File '%s' is not present in storage service, will skip.", str(path))
+        return
+
+    await data_manager.pull(path)
+    log.info("Finished pulling and extracting %s", str(path))
+
+
+def main(args=None) -> int:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument(
+        "path", help="The folder or file to get for the node", type=Path
+    )
+    options = parser.parse_args(args)
     try:
-        parser = argparse.ArgumentParser(description=__doc__)
-        parser.add_argument("path", help="The folder or file to get for the node", type=Path)
-        parser.add_argument("--silent", help="The script will silently fail if the flag is on", default=False, const=True, action="store_const", required=False)
-        options = parser.parse_args(args)
-
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(data_manager.pull(options.path))
-        return ExitCode.SUCCESS
-
-    except exceptions.S3InvalidPathError:
-        if options.silent:
-            log.warning("Could not retrieve state from S3 for %s", options.path)
-            return ExitCode.SUCCESS
-        log.exception("Could not retrieve state from S3 for %s", options.path)
+        asyncio.get_event_loop().run_until_complete(
+            pull_file_if_exists(path=options.path)
+        )
+    except Exception:  # pylint: disable=broad-except
+        log.exception(
+            "Unexpected error when retrieving state from S3 for %s", options.path
+        )
         return ExitCode.FAIL
-    except: # pylint: disable=bare-except
-        log.exception("Unexpected error when retrieving state from S3 for %s", options.path)
-        return ExitCode.FAIL
+    return ExitCode.SUCCESS
+
 
 if __name__ == "__main__":
     sys.exit(main())